fix(Routes): Refactor, fix clients path & unit tests
